fix workflow activity type label for extraction rows
