+++ src/entity-configuration/definitions/list-expanded-view-defs/processing/skeletonization-campaign.tsx
@@ -1,4 +1,4 @@
-import { TaskViewConfig } from '@/entity-configuration/definitions/list-expanded-view-defs/task-activity';
+import { TaskViewConfig } from '@/features/task';
 
 import type { ListExpandedViewConfig } from '@/entity-configuration/definitions/list-expanded-view-defs/types';
 import type { TExtendedSkeletonizationCampaignsType } from '@/entity-configuration/domain/processing/skeletonization-campaign';
